requirements.txt
fastapi[standard]==0.112.2
transformers==4.44.2
torch~=2.2.2
python-multipart
pandas~=2.2.2
openpyxl
pydantic~=2.9.0
numpy~=1.26.4
pyjwt[crypto]
sqlmodel~=0.0.22